Stop killing conky processes with pgrep and pkill (fixes #1)
Having the install script ran with `bash -c` cmd has the consequence of
breaking the installation process due to the fact that after spawning the
walle process the `pgrep -f conky` cmd in the walle script matches the
installation pid becuase `bash -c` passes the whole script as string argument.

We decided that there is not need to kill other conky processes when starting
or stopping walle because many users are running simultaneously more than one
conky instances.

Fixes #1