Load i18n texts for the choosen lang (refs #12)
The user now can choose the language in which the text literals
should appear in the conky render context via the `--lang` config
option. Each language's texts should be given as a dict file in
properties form (key=value).

Once the language is set, the lua script loads the texts corresponding
to the choosen lang.

The conky texts are organized in text lines related to a specific subject,
for instance "CPU, mem and disk" loads should be a single text line
corresponding to a single conky text line.

Refs #12