Interpolate extra vars on top of the conky vars (refs #12)
Interpolation of extra varibles is now possible in addition to the already
built-in conky variables. For instance the release name of the os is now
available in dictionary lang files under the name `$_rls_name`.

Each extra interpolation var should be given with the `$_` as prefix and
must be resolved in the main lua file. Interpolation vars are grouped to those
must be initialized once at init and to those at interval periods.

Refs #12