Optimize resolve having a global conky updates index
By having conky `updates` parsed once at the conky main call we expect
less overhead in parsing calls. Previously the code was calling the
parsing of `updates` 4 times even though this value doesn't change
within the same conky loop. So now we do it once and for all calls of
the resolve operation.