Catch and handle raised exceptions per resolving call
We better organized the code in order to catch and handle failed
resolving calls in the way to not block and swallowing calls following
a failed resolve call. Each failing call should only report the error
and return a nullish value back to its caller.

Moreover with this change we managed to enforce the public API each
resolve module should implement.