Improve logging to support debug messages
So far we were importing a new logger per executable, so both
walle and resolver script had its own logger instance. In order
to have debug mode enabled debug messages we refactor the logger
exported as a router of loggers and having an open API to update
the log level per logging route.