Format numbers so the conky lines stay fixed in length
Because dynamic numerical values like percents, temps etc. have different
number of digits over the time considering that they are flactuating or increasing
in a dynamic way, the conky text lines have to be adjusted in length each time.
This creates a bad look and feel exp, so we format each number value with
fixed number of digits and zero filled to keep line length fixed.