Fix pots coloring in the blocks component
Pots in the blocks component should be colored according to the
pot index and the current scalar value of the component. At low
scalar values first pots should be colored with a redish color,
pots in the middle should be colored with a neutral color and the
last hight index pots should get an intense yellowish color.